log: put LogTracer in its own module
